refactor(network-service): rename downstream module to store
