+++ battery-packs/network-service-battery-pack/templates/service/src/store.rs
